+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update FR translation
 * Tracking information in email reformatted
 * Tracking information moved after order items table in order email
+* Remove use of id for order,u se get_id() instead
 
 ### 1.0.0: February 2nd, 2017
 Version by Rémy Corson https://github.com/corsonr/